requirements.txt
requests>=2.31.0
deep-translator>=1.11.4
playwright>=1.40.0
python-dotenv>=1.0.0
